разобъёт код на модули и заведёт пустые файлы под будущие модули
